+++ NEWS.md
@@ -4,6 +4,7 @@
 * Added only.files parameter to get.input.files.
 * Fixed bug in the parsing of sourced scripts
 * Changed the getter functions so that when they return an empty data frame, it still has column names
+* Export ProvInfo class
 
 # provParseR 0.1.1